Move api libs. Fix svg images.
