+++ app/javascript/App/bundles/NewsfeedPage/FilterPanel/filterComponents/Coins.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import * as _ from 'lodash'
 import AsyncSelect from 'react-select/lib/Async'
 import { CoinSlug } from '~/bundles/NewsfeedPage/types'
-import localApi from '../../../../lib/localAPI'
+import localApi from '../../../common/utils/localAPI'
 
 interface Props {
   selectedCoins: CoinSlug[]
