+++ app/javascript/App/bundles/common/utils/images.ts
@@ -1,11 +1,11 @@
-export const bellIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="14" height="17" viewBox="0 0 14 17">
+export const bellIcon = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="14" height="17" viewBox="0 0 14 17">
 <g fill="none" fill-rule="evenodd">
     <path fill="#2FAEED" fill-rule="nonzero" d="M7 16.333c.917 0 1.667-.75 1.667-1.666H5.333c0 .916.75 1.666 1.667 1.666zm5-5V7.167c0-2.559-1.358-4.7-3.75-5.267v-.567C8.25.642 7.692.083 7 .083s-1.25.559-1.25 1.25V1.9C3.367 2.467 2 4.6 2 7.167v4.166L.333 13v.833h13.334V13L12 11.333zm-1.667.834H3.667v-5c0-2.067 1.258-3.75 3.333-3.75s3.333 1.683 3.333 3.75v5z"/>
     <path d="M-3-2h20v20H-3z"/>
 </g>
 </svg>`
 
-export const bulbIcon = `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="10" height="14" viewBox="0 0 10 14">
+export const bulbIcon = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="10" height="14" viewBox="0 0 10 14">
 <defs>
     <path id="a" d="M0 0h16v16H0z"/>
 </defs>
@@ -17,42 +17,42 @@ export const bulbIcon = `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="ht
 </g>
 </svg>`
 
-export const chartIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="20" height="12" viewBox="0 0 20 12">
+export const chartIcon = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="20" height="12" viewBox="0 0 20 12">
 <g fill="none" fill-rule="evenodd">
     <path fill="#2FAEED" fill-rule="nonzero" d="M14 0l2.29 2.29-4.88 4.88-4-4L0 10.59 1.41 12l6-6 4 4 6.3-6.29L20 6V0z"/>
     <path d="M-2-6h24v24H-2z"/>
 </g>
 </svg>`
 
-export const cornerArrowIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 18 18">
+export const cornerArrowIcon = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 18 18">
 <g fill="none" fill-rule="evenodd">
     <path d="M-3-3h24v24H-3z"/>
     <path fill="#2FAEED" fill-opacity=".54" fill-rule="nonzero" d="M16 16H2V2h7V0H2a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14c1.1 0 2-.9 2-2V9h-2v7zM11 0v2h3.59l-9.83 9.83 1.41 1.41L16 3.41V7h2V0h-7z"/>
 </g>
 </svg>`
 
-export const filterBtn = `<svg xmlns="http://www.w3.org/2000/svg" width="18" height="12" viewBox="0 0 18 12">
+export const filterBtn = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="18" height="12" viewBox="0 0 18 12">
 <g fill="none" fill-rule="evenodd">
     <path fill="#000" fill-opacity=".87" fill-rule="nonzero" d="M7 12h4v-2H7v2zM0 0v2h18V0H0zm3 7h12V5H3v2z"/>
     <path d="M-3-6h24v24H-3z"/>
 </g>
 </svg>`
 
-export const filterIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="18" height="12" viewBox="0 0 18 12">
+export const filterIcon = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="18" height="12" viewBox="0 0 18 12">
 <g fill="none" fill-rule="evenodd">
     <path fill="#2FAEED" fill-rule="nonzero" d="M7 12h4v-2H7v2zM0 0v2h18V0H0zm3 7h12V5H3v2z"/>
     <path d="M-3-6h24v24H-3z"/>
 </g>
 </svg>`
 
-export const listIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="18" height="12" viewBox="0 0 18 12">
+export const listIcon = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="18" height="12" viewBox="0 0 18 12">
 <g fill="none" fill-rule="evenodd">
     <path d="M-3-6h24v24H-3z"/>
     <path fill="#2FAEED" fill-rule="nonzero" d="M0 7h2V5H0v2zm0 5h2v-2H0v2zM0 2h2V0H0v2zm4 5h14V5H4v2zm0 5h14v-2H4v2zM4 0v2h14V0H4z"/>
 </g>
 </svg>`
 
-export const loading = `<svg version="1.1" id="loader-1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
+export const loading = `data:image/svg+xml;utf8,<svg version="1.1" id="loader-1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
 width="40px" height="40px" viewBox="0 0 40 40" enable-background="new 0 0 40 40" xml:space="preserve">
 <path opacity="0.2" fill="#000" d="M20.201,5.169c-8.254,0-14.946,6.692-14.946,14.946c0,8.255,6.692,14.946,14.946,14.946
 s14.946-6.691,14.946-14.946C35.146,11.861,28.455,5.169,20.201,5.169z M20.201,31.749c-6.425,0-11.634-5.208-11.634-11.634
@@ -69,21 +69,21 @@ C22.32,8.481,24.301,9.057,26.013,10.047z">
 </path>
 </svg>`
 
-export const logoReddit = `<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20">
+export const logoReddit = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20">
 <g fill="none" fill-rule="evenodd">
     <path d="M-2-2h24v24H-2z"/>
     <path fill="#FF5700" d="M12.221 13.025c.1.1.1.256 0 .346-.636.636-1.864.691-2.221.691s-1.585-.055-2.221-.691c-.1-.09-.1-.246 0-.346.09-.09.246-.09.335 0 .402.412 1.272.546 1.886.546.614 0 1.473-.133 1.886-.546.09-.09.246-.09.335 0zm-3.426-2.02a.996.996 0 0 1-.994.993.998.998 0 1 1 0-1.998c.547 0 .994.446.994 1.004zm4.408 0a.998.998 0 0 1-1.004.993.996.996 0 0 1-.994-.994.998.998 0 1 1 1.998 0zm2.801-1.34c0-.736-.602-1.328-1.339-1.328-.38 0-.714.156-.96.402-.904-.625-2.12-1.027-3.47-1.072l.703-3.158 2.232.502c0 .547.446.993.993.993.558 0 1.004-.457 1.004-1.004 0-.547-.446-1.004-1.004-1.004-.39 0-.726.234-.893.558l-2.466-.547c-.123-.034-.246.056-.28.178l-.77 3.482c-1.339.056-2.544.458-3.448 1.083a1.34 1.34 0 0 0-.971-.413c-.737 0-1.34.592-1.34 1.328 0 .536.313.982.76 1.206-.045.2-.067.412-.067.625 0 2.12 2.388 3.839 5.323 3.839 2.947 0 5.335-1.719 5.335-3.84 0-.212-.022-.435-.078-.636a1.34 1.34 0 0 0 .736-1.194zM20 10c0 5.525-4.475 10-10 10S0 15.525 0 10 4.475 0 10 0s10 4.475 10 10z"/>
 </g>
 </svg>`
 
-export const logoTwitter = `<svg xmlns="http://www.w3.org/2000/svg" width="19" height="15" viewBox="0 0 19 15">
+export const logoTwitter = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="19" height="15" viewBox="0 0 19 15">
 <g fill="none" fill-rule="evenodd">
     <path fill="#2FAEED" d="M18.438 1.797c-.47.729-1.094 1.38-1.875 1.953v.469c0 2.734-.984 5.215-2.95 7.441C11.647 13.887 9.036 15 5.781 15c-2.083 0-4.01-.573-5.781-1.719.417.052.716.078.898.078 1.745 0 3.308-.546 4.688-1.64a3.803 3.803 0 0 1-2.188-.762A3.578 3.578 0 0 1 2.07 9.102a4.7 4.7 0 0 0 .703.078c.26 0 .6-.052 1.016-.157a3.818 3.818 0 0 1-2.168-1.308 3.6 3.6 0 0 1-.879-2.402v-.04c.573.313 1.146.47 1.719.47C1.34 5.012.78 3.957.78 2.577c0-.65.17-1.276.508-1.875 2.005 2.474 4.596 3.79 7.774 3.945-.053-.182-.079-.468-.079-.859 0-1.068.365-1.966 1.094-2.695C10.808.364 11.706 0 12.773 0c1.068 0 1.98.404 2.735 1.21.781-.13 1.588-.442 2.422-.937-.26.912-.82 1.615-1.68 2.11a6.451 6.451 0 0 0 2.188-.586z"/>
     <path d="M-3-5h24v24H-3z"/>
 </g>
 </svg>`
 
-export const watchListStarIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="50" height="50" viewBox="0 0 50 50">
+export const watchListStarIcon = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="50" height="50" viewBox="0 0 50 50">
 <g fill="none" fill-rule="evenodd">
     <path stroke="#000" stroke-dasharray="6,2" stroke-opacity=".38" stroke-width="1.5" d="M25.186 42.03l-12.631 6.953a2 2 0 0 1-2.938-2.076l2.435-14.867-10.4-10.615a2 2 0 0 1 1.128-3.377l14.289-2.173 6.307-13.382a2 2 0 0 1 3.619 0l6.307 13.382 14.289 2.173a2 2 0 0 1 1.128 3.377l-10.4 10.615 2.435 14.867a2 2 0 0 1-2.938 2.076l-12.63-6.953z"/>
     <path d="M-3-2h56v56H-3z"/>
