+++ app/javascript/App/components/CoinIndex/index.js
@@ -2,7 +2,7 @@ import React, { Component, Fragment } from 'react'
 import { Table } from 'antd'
 import ColumnNames from './ColumnNames'
 import SearchCoins from '~/bundles/common/components/SearchCoins'
-import API from '../../lib/localAPI'
+import API from '../../bundles/common/utils/localAPI'
 import * as _ from 'lodash'
 
 class CoinIndex extends Component {
