+++ app/javascript/App/lib/misc.js
@@ -1,3 +1,4 @@
+//TODO: Remove - Redux only
 import * as _ from 'lodash'
 import inflection from 'lodash-inflection'
 _.mixin(inflection)
