+++ app/javascript/App/lib/stateHelpers.js
@@ -1,3 +1,4 @@
+//TODO: Remove - Redux only
 export const listIndex = (list, value, key = 'key') => {
   let index = list.size
   const existingIndex = list.findIndex((o) => o.get(key) === value)
