+++ package.json
@@ -79,7 +79,9 @@
     "@storybook/addon-links": "^3.4.8",
     "@storybook/addons": "^3.4.8",
     "@storybook/react": "^3.4.8",
+    "@types/axios": "^0.14.0",
     "@types/lodash": "^4.14.116",
+    "@types/qs": "^6.5.1",
     "@types/react": "^16.4.9",
     "@types/react-currency-formatter": "^1.1.0",
     "@types/react-dom": "^16.0.7",
